--- src/scripts/priorityLevel.js
@@ -1,11 +1,0 @@
-// const listItem = document.querySelector(".task-list-neutral")
-
-// const addPriorityColors = (priorityLevel) => {
-//     if (priorityLevel === 1) {
-//         listItem.classList.add("leastPriority")
-//     } else if (priorityLevel === 2) {
-//         listItem.classList.add("middlePriority")
-//     } else {
-//         listItem.classList.add("firstPriority")
-//     }
-// }
