--- src/scripts/taskCreateComponent.js
@@ -1,9 +1,0 @@
-/*
-    Author: Emily
-    Name: taskCreateComponent
-    Purpose: Displaying the "New Task" button, and attaching an event listener that presents the "New Task" form when clicked.
-*/
-
-
-
-
