--- src/scripts/taskEventHandlers.js
@@ -1,15 +1,0 @@
-// import API from "./taskApiManager"
-
-// const createTaskHandlers = {
-//     completeTask: function () {
-//         API.editTaskData()
-//         .then(parsedResult  => {
-//             parsedResult.taskCompleted === true;
-//         })
-//     }
-// }
-
-// export default createTaskHandlers
-// const btn = document.querySelector(".btn btn-outline-secondary task-complete-btn");
-// btn.addEventListener("click", createTaskHandlers.completeTask())
-// export default createTaskHandlers
